+++ db4e/Constants/DFile.py
@@ -18,6 +18,7 @@ class DFile(ConstGroup):
     CHOWN : str = "chown"
     CONFIG_FILE : str = DField.CONFIG_FILE
     LOGROTATE : str = "logrotate"
+    MONGODUMP : str = "mongodump"
     P2POOL_LOG : str = "p2pool.log"
     P2POOL_STDIN : str = "p2pool.stdin"
     STATS_MOD : str = "stats_mod"